added ability to manage skills through PATHC + test

+++ server/src/modules/users/users.service.ts
@@ -3,8 +3,9 @@ import { GetUsersDTO, LoginDTO, UpdateMeDTO } from "./users.dto";
 import jwt from 'jsonwebtoken'
 import { User } from "./users.model";
 import { config } from "../../config/config";
-import UserMapper from "./users.mapper";
+import UserMapper from "./users.mappers";
 import deleteFile from "../../common/utils/utils.deleteFile";
+import { Skill } from "../skills/skills.model";
 
 class UsersService {
 	static async isUserValid(dto: LoginDTO) {
@@ -24,16 +25,21 @@ class UsersService {
 	static async updateMe(userId: string, dto: UpdateMeDTO) {
 		const user = await User.findById(userId)
 		if (!user) throw new ErrorWithStatus(404, "User not found")
+		if (dto.skills) {
+			const skills = await Skill.find({ _id: { $in: dto.skills } })
+			if (skills.length !== dto.skills.length) throw new ErrorWithStatus(400, "One or more skill IDs are invalid")
+		}
 		if (dto.photo_path && user.photo_path) {
 			await deleteFile(user.photo_path)
 		}
 		user.set(dto)
 		await user.save()
-		return UserMapper.toPublicUser(user)
+		const updatedUser = await User.findById(userId).populate('skills');
+		return UserMapper.toPublicUser(updatedUser!)
 	}
 
 	static async getUser(userId: string) {
-		const user = await User.findById(userId)
+		const user = await User.findById(userId).populate('skills');
 		if (!user) throw new ErrorWithStatus(404, "User not found")
 		return UserMapper.toPublicUser(user)
 	}
@@ -51,7 +57,7 @@ class UsersService {
 				{ last_name: { $regex: dto.nameContains, $options: 'i' } }
 			]
 		}
-		const users = await User.find(query).skip(dto.offset).limit(dto.limit)	
+		const users = await User.find(query).skip(dto.offset).limit(dto.limit).populate('skills')	
 		return users.map(user => UserMapper.toPublicUser(user))
 	}
 }
